04.04.2024 Update

+++ postinstall.sh
@@ -26,9 +26,11 @@ _conf() {
     sed -i "s/Session=.*/Session=plasmax11/" /etc/sddm.conf
   fi
 
-  echo "QT_QPA_PLATFORMTHEME=qt5ct" > /etc/environment
-  #echo "QT_STYLE_OVERRIDE=kvantum" > /etc/environment
-  echo "GTK_THEME=Ctlos-Dark" >> /etc/environment
+  if [ ! -f "/usr/share/xsessions/plasmax11.desktop" ]; then
+    echo "QT_QPA_PLATFORMTHEME=qt5ct" > /etc/environment
+    #echo "QT_STYLE_OVERRIDE=kvantum" > /etc/environment
+    echo "GTK_THEME=Ctlos-Dark" >> /etc/environment
+  fi
 }
 
 clean_post
